requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
beautifulsoup4==4.12.3
PyJWT==2.8.0
python-multipart==0.0.9
